Initial commit.
I copied the configuration of Mezz_Foundation and I adjusted it so all instances of "Foundation"
said "PackageName" instead and tried to make it as suitable as possible for reuse.

I check that the sample builds and the tests run and pass. Even though the tests are trivial, it
is useful to know that they are easy to set up for new users.